adding error handling

+++ src/main/java/com/mercadolibre/apicompliance/repository/ProcessRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProcessRepository extends JpaRepository<Process,Integer> {
-    List<Process> getAllByAuditId(Integer id);
+public interface ProcessRepository extends JpaRepository<Process,Long> {
+    List<Process> getAllByAuditId(Long id);
 }
